+++ CHANGES.md
@@ -1,3 +1,11 @@
+## Unreleased
+
+### Added
+
+- **irmin**
+  - Add `test_set_and_get*` functions to retrieve commit associated with an update to the store (#2075, @patricoferris)
+
+
 ## 3.4.1 (2022-09-07)
 
 ### Added